services/Manager/requirements.txt
click==7.1.2
coloredlogs==14.0
Flask==1.1.2
Flask-HTTPAuth==4.1.0
gitdb==4.0.5
GitPython==3.1.7
gunicorn==20.0.4
humanfriendly==8.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.0
pandas==1.0.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
smmap==3.0.4
Werkzeug==1.0.1
gunicorn==20.0.4
console-menu